Dynamically render list of games based on date input and update UI color theme
This commit creates a custom hook that fetches from the MLB Stats API to dynamically render a list of all games based on the selected date. Entered dates that go beyond the input's min/max values are capped at the min/max value, preventing users from entering invalid dates.

This commit also updates the UI theme by changing the colors of table headers and the footer from blue to gray. This prevents the primary CTA button frmo competing with the rest of the app visually, helping it stick out to users as an actionable element.