Implement dynamic data rendering in GeneralTabContent
This commit updates the GameResponse type to better reflect the shape of the response in case of missing/undefined values. When rendering dynamic content in the general tab section, utility functions were created to help format and render the correct content in the UI as well as keep GeneralTabContent clean.